Extract status api endpoints into separate file
